fix bugs
FE user:
- validate các field user, total, address ở server khi checkout
- Fix signup validation (bằng bootstrap)
FE quản trị:
- fix update, create voucher (bug là do fetchtype.all)
- authorize: ẩn các btn nếu role là STAFF
- quản trị user: fix STAFF ko tạo mới đc ADMIN, ko tự xóa bản thân , ko đc xóa ADMIN
- fix layout các list.html btn delete và edit

+++ src/main/java/com/edu/graduationproject/controller/rest/UserRestController.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -61,7 +62,7 @@ public class UserRestController {
             map.put("image_url", loggedinUser.get().getImage_url());
             map.put("roles", userRoles);
             return ResponseEntity.ok(map);
-        } catch (NoSuchElementException e) {
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
@@ -70,11 +71,22 @@ public class UserRestController {
     public ResponseEntity<User> getUserByUsername(@PathVariable("idOrUsername") Optional<Object> idOrUsername) {
         try {
             Optional<User> user = userService.findByUsername((String) idOrUsername.get());
+            boolean isExist = false;
             if (!user.isPresent()) {
-                user = userService.findById(Integer.valueOf((String) idOrUsername.get()));
+                user = userService.findById(Integer.parseInt((String) idOrUsername.get()));
+                if (user.isPresent()) {
+                    if (user.get().getIsDeleted() == false) {
+                        isExist = true;
+                    }
+                }
+            } else {
+                isExist = true;
             }
-            return ResponseEntity.ok(user.get());
-        } catch (NoSuchElementException e) {
+            if (isExist == true) {
+                return ResponseEntity.ok(user.get());
+            }
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
@@ -82,6 +94,16 @@ public class UserRestController {
     // admin
     @PostMapping("/rest/users")
     public ResponseEntity<User> create(@RequestBody User user) {
+        Optional<User> findUser = userService.findByUsername(user.getUsername());
+        boolean isExist = false;
+        if (findUser.isPresent()) {
+            if (findUser.get().getIsDeleted() == false) {
+                isExist = true;
+            }
+        }
+        if (isExist == true) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(userService.save(user));
     }
 
@@ -100,7 +122,7 @@ public class UserRestController {
                 return ResponseEntity.ok(savedUser);
             }
             return ResponseEntity.notFound().build();
-        } catch (NoSuchElementException e) {
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
@@ -120,7 +142,7 @@ public class UserRestController {
                 return ResponseEntity.ok(savedUser);
             }
             return ResponseEntity.notFound().build();
-        } catch (NoSuchElementException e) {
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
@@ -131,14 +153,30 @@ public class UserRestController {
         try {
             Optional<User> reqUser = userService.findByUsername(auth.getName());
             Optional<User> findUser = userService.findByUsername(username.get());
+            boolean isExist = false;
             if (findUser.isPresent() && reqUser.isPresent()) {
+                isExist = true;
+            }
+            if (isExist == true) {
                 if (auth.getName().equals(username.get())) {
                     throw new BadRequestException("You can't delete yourself");
                 }
-                List<UserRole> reqRoles = userRoleService.findRolesOfAdministrators();
+                boolean isReqUserStaff = auth.getAuthorities().stream()
+                        .map(role -> role.getAuthority().replace("ROLE_", ""))
+                        .collect(Collectors.toList())
+                        .contains("STAFF");
+                if (isReqUserStaff) {
+                    boolean isFindUserAdmin = findUser.get().getAuthorities().stream()
+                            .map(role -> role.getRole().getName())
+                            .collect(Collectors.toList()).contains("ADMIN");
+                    if (isFindUserAdmin) {
+                        throw new BadRequestException("You can't delete admin");
+                    }
+                }
+                userService.deleteByUsername(username.get());
+            } else {
+                throw new ResourceNotFoundException("User", "username", username.get());
             }
-            userService.deleteByUsername(username.get());
-            throw new ResourceNotFoundException("User", "username", username.get());
         } catch (Exception e) {
             throw new ResourceNotFoundException("User", "username", username.get());
         }
